# Conflicts:
#	src/main/java/com/azya/partstore/services/PartServiceImpl.java

+++ src/main/java/com/azya/partstore/services/PartService.java
@@ -11,6 +11,6 @@ public interface PartService {
 
     void deletePart(Part part);
 
-    int getAvailableComputersCount();
+    long getAvailableComputersCount();
 
 }
